+++ daiv/daiv/settings/components/common.py
@@ -1,3 +1,5 @@
+from pathlib import Path
+
 from decouple import Csv, config
 from get_docker_secret import get_docker_secret
 
@@ -11,24 +13,58 @@ SITE_ID = 1
 
 LOCAL_APPS = ["accounts", "automation", "codebase", "core", "slash_commands"]
 
-THIRD_PARTY_APPS = ["crontask", "django_extensions", "django_tasks", "django_tasks_db"]
+THIRD_PARTY_APPS = [
+    "crontask",
+    "django_extensions",
+    "django_tasks",
+    "django_tasks_db",
+    "allauth",
+    "allauth.account",
+    "allauth.socialaccount",
+    "allauth.socialaccount.providers.github",
+    "allauth.socialaccount.providers.gitlab",
+]
 
-DJANGO_APPS = ["django.contrib.auth", "django.contrib.contenttypes", "django.contrib.sessions"]
+DJANGO_APPS = [
+    "django.contrib.auth",
+    "django.contrib.contenttypes",
+    "django.contrib.sessions",
+    "django.contrib.sites",
+    "django.contrib.messages",
+    "django.contrib.staticfiles",
+]
 
-# See: https://docs.djangoproject.com/en/dev/ref/settings/#installed-apps
 INSTALLED_APPS = LOCAL_APPS + THIRD_PARTY_APPS + DJANGO_APPS
 
 
 MIDDLEWARE = [
     "django.middleware.security.SecurityMiddleware",
+    "whitenoise.middleware.WhiteNoiseMiddleware",
     "django.contrib.sessions.middleware.SessionMiddleware",
     "django.middleware.common.CommonMiddleware",
+    "django.middleware.csrf.CsrfViewMiddleware",
+    "django.contrib.auth.middleware.AuthenticationMiddleware",
+    "django.contrib.messages.middleware.MessageMiddleware",
+    "allauth.account.middleware.AccountMiddleware",
 ]
 
 
 # TEMPLATE CONFIGURATION - https://docs.djangoproject.com/en/dev/ref/settings/#templates
 
-TEMPLATES = [{"BACKEND": "django.template.backends.django.DjangoTemplates", "APP_DIRS": True}]
+TEMPLATES = [
+    {
+        "BACKEND": "django.template.backends.django.DjangoTemplates",
+        "APP_DIRS": True,
+        "OPTIONS": {
+            "context_processors": [
+                "django.template.context_processors.debug",
+                "django.template.context_processors.request",
+                "django.contrib.auth.context_processors.auth",
+                "django.contrib.messages.context_processors.messages",
+            ]
+        },
+    }
+]
 
 DEFAULT_AUTO_FIELD = "django.db.models.BigAutoField"
 
@@ -38,7 +74,6 @@ WSGI_APPLICATION = "daiv.wsgi.application"
 
 
 # Password validation
-# https://docs.djangoproject.com/en/1.9/ref/settings/#auth-password-validators
 
 AUTH_PASSWORD_VALIDATORS = [
     {"NAME": "django.contrib.auth.password_validation.UserAttributeSimilarityValidator"},
@@ -76,3 +111,16 @@ SECURE_REFERRER_POLICY = "strict-origin-when-cross-origin"
 # Use CSRF in session instead of cookie: https://docs.djangoproject.com/en/dev/ref/csrf/
 CSRF_USE_SESSIONS = True
 X_FRAME_OPTIONS = "DENY"
+
+
+# STATIC FILES - https://docs.djangoproject.com/en/dev/ref/settings/#static-files
+
+STATIC_URL = "/static/"
+STATIC_ROOT = Path.home() / "data" / "static"
+STORAGES = {"staticfiles": {"BACKEND": "whitenoise.storage.CompressedManifestStaticFilesStorage"}}
+
+
+# EMAIL
+
+EMAIL_BACKEND = config("EMAIL_BACKEND", default="django.core.mail.backends.console.EmailBackend")
+DEFAULT_FROM_EMAIL = config("DEFAULT_FROM_EMAIL", default="noreply@daiv.dev")